tile placement updates can now be subscribed to abritraryly. AOE buffs for tech buildings partially implemented

+++ Assets/Scripts/Map/Tiles/MappedTiles/GeothermalVentTile.cs
@@ -87,6 +87,7 @@ public class GeothermalVentShellTile : ResourceTile
 	public override void Destroy()
 	{
 		base.Destroy();
-		Map.EM.DestroyEntity(_shell);
+		if(Map.EM.Exists(_shell))
+			Map.EM.DestroyEntity(_shell);
 	}
 }
